Fix maven modules, Add base module to handle basic spigot stuff instead of having an API with a spigot dependency
