+++ .idea/compiler.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="CompilerConfiguration">
+    <annotationProcessing>
+      <profile name="Maven default annotation processors profile" enabled="true">
+        <sourceOutputDir name="target/generated-sources/annotations" />
+        <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
+        <outputRelativeToContentRoot value="true" />
+        <module name="elitenpcs-api" />
+        <module name="elitenpcs-spigot" />
+        <module name="v1_8_R3" />
+      </profile>
+    </annotationProcessing>
+    <bytecodeTargetLevel>
+      <module name="EliteNPCs" target="1.5" />
+      <module name="elitenpcs-api" target="1.5" />
+      <module name="elitenpcs-spigot" target="1.5" />
+      <module name="v1_8_R3" target="1.5" />
+    </bytecodeTargetLevel>
+  </component>
+</project>
