+++ .idea/modules.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/EliteNPCs.iml" filepath="$PROJECT_DIR$/EliteNPCs.iml" />
+      <module fileurl="file://$PROJECT_DIR$/elitenpcs-api/elitenpcs-api.iml" filepath="$PROJECT_DIR$/elitenpcs-api/elitenpcs-api.iml" />
+      <module fileurl="file://$PROJECT_DIR$/elitenpcs-spigot/elitenpcs-spigot.iml" filepath="$PROJECT_DIR$/elitenpcs-spigot/elitenpcs-spigot.iml" />
+      <module fileurl="file://$PROJECT_DIR$/v1_8_R3/v1_8_R3.iml" filepath="$PROJECT_DIR$/v1_8_R3/v1_8_R3.iml" />
+    </modules>
+  </component>
+</project>
