+++ elitenpcs-api/pom.xml
@@ -2,10 +2,13 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>EliteNPCs</artifactId>
+        <groupId>io.github.elitedev</groupId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>io.github.elitedev</groupId>
-    <artifactId>elitenpcs-v1_8_R3</artifactId>
-    <version>1.0-SNAPSHOT</version>
+    <artifactId>elitenpcs-api</artifactId>
 
 </project>
