+++ elitenpcs-api/src/main/java/io/github/elitedev/entity/NPC.java
@@ -1,0 +1,13 @@
+package io.github.elitedev.entity;
+
+import java.util.UUID;
+
+public interface NPC {
+
+  UUID getUniqueId();
+
+  String getName();
+
+  void setName(String name);
+
+}
